# p5j/forest-02 Optimize via variable caching and use of Quadtree