--- a/p5js/common/examples/quadtree-2/index.md
+++ b/p5js/common/examples/quadtree-2/index.md
@@ -7,6 +7,7 @@ excerpt: Implementation of a quadtree data structure, which groups data points i
 
 js_scripts:
 - https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.7.2/p5.js
+- /sketchbook/vendor/dat.gui/0.7.9/dat.gui.min.js
 - /sketchbook/p5js/common/p5js_settings.js
 - /sketchbook/p5js/common/p5js_utils.js
 - /sketchbook/js/util_functions.js