--- a/p5js/common/examples/quadtree-2/local_dev.html
+++ b/p5js/common/examples/quadtree-2/local_dev.html
@@ -6,6 +6,7 @@
 
     <script src="//cdnjs.cloudflare.com/ajax/libs/inobounce/0.2.1/inobounce.js" charset="utf-8"></script>
     <script src="/sketchbook/vendor/p5.js/0.7.2/p5.js" charset="utf-8"></script>
+    <script src="/sketchbook/vendor/dat.gui/0.7.9/dat.gui.min.js"></script>
     <script src="/sketchbook/p5js/common/p5js_settings.js" charset="utf-8"></script>
     <script src="/sketchbook/p5js/common/p5js_utils.js" charset="utf-8"></script>
     <script src="/sketchbook/js/util_functions.js" charset="utf-8"></script>