--- a/p5js/forest-02/classes/forest.js
+++ b/p5js/forest-02/classes/forest.js
@@ -35,17 +35,28 @@ class Forest {
 
   sproutTreeForSpecies(x, y, species){
     // TODO: Remove this cap, or make more prominent
-    if (this.trees.length > 700) { return; }
+    if (this.trees.length > 700) { 
+      // consol/e.log('no more room at the inn');
+      return; 
+    }
     this.trees.push( new Tree(x, y, 0, this.treeCounter++, species) );
   }
 
   tick(){
+    this.rebuildQuadSearchTree();
     this.trees.forEach(t => t.tick( this.resourcesForTree(t) ));
     this.lossDueToCompetition();
     this.lossDueToMaxAge();
     this.lossDueToForaging();
     this.triggerSeasonalBehavior();
     this.lossDueToOutOfBounds();
+  }
+
+  rebuildQuadSearchTree(){
+    this.quadtree = new Quadtree(this.area, 10, false);
+    for (let i = 0; i < this.trees.length; i++){
+      this.quadtree.add(this.trees[i]);
+    }
   }
 
   lossDueToCompetition(){
@@ -75,9 +86,10 @@ class Forest {
   }
 
   resourcesForTree(tree){
-    // TODO: Use QuadTree to efficiently isolate trees in competition
     // TODO: Longer term, consider beneficial aspects of mother trees, and same species
-    return this.trees.filter(t => t != tree)
+    const competingTrees = this.quadtree.find(tree);
+    
+    return competingTrees.filter(t => t != tree)
                      .map(otherTree => this.resourcesClaimedByTree(otherTree, tree))
                      .reduce((remainder, val) => remainder - val, 1);
   }
@@ -110,7 +122,7 @@ class Forest {
     }
 
     const rootCompFactor = 0.5;
-    const sunlightLossFactor = (losingTree.height < claimingTree.height) ? 0.5 : 0;
+    const sunlightLossFactor = (losingTree.verticalExtent < claimingTree.verticalExtent) ? 0.5 : 0;
     const combinedLossFactor = (rootCompFactor + sunlightLossFactor);
 
     // % Resource Loss = (% of Smaller Overlapping x Area of Smaller) / Area of Larger x Weighting Factors