--- a/p5js/forest-02/index.md
+++ b/p5js/forest-02/index.md
@@ -13,6 +13,7 @@ js_scripts:
 - /sketchbook/js/util_functions.js
 - /sketchbook/js/options_set.js
 - /sketchbook/js/models/rect.js
+- /sketchbook/js/models/quadtree.js
 - /sketchbook/js/models/cell_grid.js
 - classes/cell.js
 - classes/cell_viewer.js