--- a/p5js/forest-02/local_dev.html
+++ b/p5js/forest-02/local_dev.html
@@ -9,6 +9,7 @@
     <script src="/sketchbook/js/util_functions.js"></script>
     <script src="/sketchbook/js/options_set.js"></script>
     <script src="/sketchbook/js/models/rect.js"></script>
+    <script src="/sketchbook/js/models/quadtree.js"></script>
     <script src="/sketchbook/js/models/cell_grid.js" charset="utf-8"></script>
 
     <script src="classes/cell.js"></script>