Main commit: https://github.com/brianhonohan/sketchbook/commit/756e7f54af9920a3739602f4403ca62e5f626224

This had significant overall improvement
- reduce time per frame from 39ms => 11 ms
- will be able to scale up the number of trees as well

Cache the shadowRadius rather than recompute on each request
- this cut the time-per frame by more than half
- from 39 ms ==> 17ms

Use a Quadtree to reduce the number of 'dist' (square root) calcs necessary for checking if two trees overlap.

Had to refactor name of tree.height to tree.verticalExtent
- To allow for using 'height' as the width/height (on 2D screen) of where the area the tree occupies;
- Thus allow Tree objects to be placed directly into Quadtree (2D search structure)


Made possible by: https://github.com/brianhonohan/sketchbook/commit/ea10d0bc8d1325a9e358c9528fef95a57f60fa13 
- which allows for querying a `Quadtree` via a 'rect' like object
- updated example: https://brianhonohan.com/sketchbook/p5js/common/examples/quadtree-2/


----

### Perf details


* Before ... ~ 39 ms per frame

![01--PreOpt--Limit-500---NoQuadTree--RecalcShadowRadius--Screenshot 2025-02-11 225011](https://github.com/user-attachments/assets/d3f21a91-6565-4c84-a108-e9a187f2c2ee)


* Optimization 1
- Refactor `Tree.shadowRadius()` to return a computed value (that is recomputed once per frame)
- reduced frame time from 39ms to 16.5 ms

![02--Opt--Limit-500---NoQuadTree--CacheShadowRadius--Screenshot 2025-02-11 225533](https://github.com/user-attachments/assets/1993f1d3-abfb-4ac9-b763-4d6a1d908754)


* Optimization 2
- Use quadtree to look up competing trees
- this reduces the number of `dist()` calls that are made, reducing from O(n^2) to not that.
- further reduces frame time from 16.5ms to ~11 ms 
- 
![03--Opt--Limit-500---QuadTree--CacheShadowRadius--Screenshot 2025-02-11 225743](https://github.com/user-attachments/assets/46c81fd0-3042-4f32-a9da-2a5af14e480f)